refactor: TeamChecklistFragmentAdapter 클래스 분리
