+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/team/TeamChecklistActivity.kt
@@ -3,23 +3,25 @@ package com.bottari.presentation.view.checklist.team
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.bottari.presentation.R
 import com.bottari.presentation.common.base.BaseActivity
 import com.bottari.presentation.databinding.ActivityTeamChecklistBinding
-import com.bottari.presentation.view.checklist.team.checklist.TeamChecklistFragment
+import com.bottari.presentation.view.checklist.team.checklist.TeamChecklistFragmentAdapter
 import com.google.android.material.tabs.TabLayoutMediator
 
 class TeamChecklistActivity : BaseActivity<ActivityTeamChecklistBinding>(ActivityTeamChecklistBinding::inflate) {
+    val bottariId: Long by lazy { intent.getLongExtra(EXTRA_BOTTARI_ID, INVALID_BOTTARI_ID) }
+
+    val adapter: TeamChecklistFragmentAdapter by lazy {
+        TeamChecklistFragmentAdapter(this, bottariId)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setupUI()
     }
 
     private fun setupUI() {
-        val bottariId = intent.getLongExtra(EXTRA_BOTTARI_ID, INVALID_BOTTARI_ID)
         val bottariTitle = intent.getStringExtra(EXTRA_BOTTARI_TITLE)
         val adapter = TeamChecklistFragmentAdapter(this, bottariId)
         binding.tvBottariTitle.text = bottariTitle
@@ -33,20 +35,6 @@ class TeamChecklistActivity : BaseActivity<ActivityTeamChecklistBinding>(Activit
                     else -> getString(R.string.tap_member_checklist)
                 }
         }.attach()
-    }
-
-    private class TeamChecklistFragmentAdapter(
-        fa: FragmentActivity,
-        private val bottariId: Long,
-    ) : FragmentStateAdapter(fa) {
-        override fun getItemCount(): Int = 3
-
-        override fun createFragment(position: Int): Fragment =
-            when (position) {
-                0 -> TeamChecklistFragment.newInstance(bottariId)
-                1 -> TeamStatusFragment.newInstance()
-                else -> MemberStatusFragment.newInstance()
-            }
     }
 
     companion object {
